requirements.txt
numpy
imutils
sklearn
tensorflow
keras
opencv2-python
